requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
mysql-connector-python==9.4.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0

# Install these into your virtualenv with:
# /home/charlesbr101/V-Lab_Internship/.venv/bin/python -m pip install -r requirements.txt
